+++ src/emulator/controller.ts
@@ -41,6 +41,7 @@ import { FLAG_EXPORT_ON_EXIT_NAME } from "./commandUtils";
 import { fileExistsSync } from "../fsutils";
 import { StorageEmulator } from "./storage";
 import { getDefaultDatabaseInstance } from "../getDefaultDatabaseInstance";
+import { getProjectDefaultAccount } from "../auth";
 
 async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Address> {
   let host = Constants.normalizeHost(
@@ -427,9 +428,11 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
       );
     }
 
+    const account = getProjectDefaultAccount(options.projectRoot);
     const functionsEmulator = new FunctionsEmulator({
       projectId,
       functionsDir,
+      account,
       host: functionsAddr.host,
       port: functionsAddr.port,
       debugPort: inspectFunctions,