+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 - Adds retries on Quota Exceeded errors during functions deployment (#2606, #1372).
 - Update to Cloud Firestore emulator v1.11.13 which includes a bug fix for documents which are created and deleted in a single transaction.
 - Fixes an edge case with nextPageToken in batchGet in Auth Emulator (#3231).
+- Removes unused dependencies (#3252).
+- Adds support for multiple accounts via new commands `login:use`, `login:add` and `login:list`.